enhanced headering

+++ Common/Src/MessageFormat.cpp
@@ -1,10 +1,10 @@
 #include <iostream>
 using namespace std;
 
-uint16_t calculateChecksum(const string& data) {
-    uint16_t checksum = 0;
+uint32_t calculateChecksum(const string& data) {
+    uint32_t checksum = 0;
     for (char ch : data) {
-        checksum += static_cast<uint16_t>(ch);
+        checksum += static_cast<uint32_t>(ch);
     }
     return checksum;
 }
